Update README.md
No clue how to comment, so I have suggested changes and comments start with Remya.

No need to accept any of my suggestions. I do no wish to screw up your repository. 

For the purpose of the overall rebuttal repo, I need the methods/descriptions, scripts and files that you used to generate path_lenght_df.csv. Perhaps in a separate folder. Nothing else. I think that is all for the main text.

If we use the euclidean distances for each pair of syllables in a reply to the reviewer's comments, this procedure could  be described, but ideally in a different folder.

Similarly, if we use other data from your UMAP exercise, we can then link to specific folders.

I hope, I am making some sense